desktop-app-v2/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.1.0
pydantic>=2.0.0
